Issue #752 Browse BAM coverage intervals - fix remarks
