+++ server/catgenome/profiles/jar/catgenome.properties
@@ -62,3 +62,12 @@ taxonomy.top.hits=${TAXONOMY_TOP_HITS:10}
 
 #HOMOLOGENE
 homologene.index.directory=${HOMOLOGENE_INDEX_DIR:./contents/homologene}
+
+#METABOLIC PATHWAY
+pathway.index.directory=${PATHWAY_INDEX_DIRECTORY:@rootDirPath@/contents/pathway}
+pathway.top.hits=${PATHWAY_TOP_HITS:100}
+
+#COVERAGE
+bam.coverage.index.directory=${BAM_COVERAGE_INDEX_DIRECTORY:@rootDirPath@/contents/coverage}
+coverage.top.hits=${COVERAGE_TOP_HITS:100000}
+coverage.batch.size=${COVERAGE_BATCH_SIZE:500}
