+++ server/catgenome/src/main/java/com/epam/catgenome/manager/bam/BamCoverageManager.java
@@ -283,6 +283,7 @@ public class BamCoverageManager {
                     coverageAreas = new ArrayList<>();
                 }
             }
+            write(coverageAreas);
             bamCoverage.setCoverage(totalCoverage / intervals);
         }
     }
