+++ server/catgenome/src/main/java/com/epam/catgenome/manager/bam/BamSecurityService.java
@@ -26,24 +26,17 @@ package com.epam.catgenome.manager.bam;
 
 import com.epam.catgenome.controller.vo.ReadQuery;
 import com.epam.catgenome.controller.vo.registration.IndexedFileRegistrationRequest;
-import com.epam.catgenome.entity.bam.BamCoverage;
 import com.epam.catgenome.entity.bam.BamFile;
 import com.epam.catgenome.entity.bam.BamQueryOption;
-import com.epam.catgenome.entity.bam.CoverageInterval;
-import com.epam.catgenome.entity.bam.CoverageQueryParams;
 import com.epam.catgenome.entity.bam.Read;
 import com.epam.catgenome.entity.reference.Sequence;
 import com.epam.catgenome.entity.track.Track;
-import com.epam.catgenome.util.db.Page;
 import lombok.RequiredArgsConstructor;
-import org.apache.lucene.queryparser.classic.ParseException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseBodyEmitter;
 
 import java.io.IOException;
-import java.util.List;
-import java.util.Set;
 
 import static com.epam.catgenome.security.acl.SecurityExpressions.*;
 
@@ -60,7 +53,6 @@ public class BamSecurityService {
                     "#query.projectId, 'READ')";
 
     private final BamManager bamManager;
-    private final BamCoverageManager coverageManager;
 
     @PreAuthorize(ROLE_ADMIN + OR + ROLE_BAM_MANAGER)
     public BamFile registerBam(IndexedFileRegistrationRequest request) throws IOException {
@@ -92,30 +84,5 @@ public class BamSecurityService {
     public void sendBamTrackToEmitterFromUrl(Track<Read> track, BamQueryOption option, String fileUrl, String indexUrl,
                                              ResponseBodyEmitter emitter) throws IOException {
         bamManager.sendBamTrackToEmitterFromUrl(track, option, fileUrl, indexUrl, emitter);
-    }
-
-    @PreAuthorize(ROLE_ADMIN + OR + ROLE_BAM_MANAGER)
-    public BamCoverage createCoverage(final BamCoverage coverage) throws IOException {
-        return coverageManager.create(coverage);
-    }
-
-    @PreAuthorize(ROLE_USER)
-    public List<BamCoverage> loadAll() throws IOException {
-        return coverageManager.loadAll();
-    }
-
-    @PreAuthorize(ROLE_USER)
-    public List<BamCoverage> loadByBamId(final Set<Long> bamIds) throws IOException {
-        return coverageManager.loadByBamId(bamIds);
-    }
-
-    @PreAuthorize(ROLE_USER)
-    public Page<CoverageInterval> loadCoverage(final CoverageQueryParams params) throws ParseException, IOException {
-        return coverageManager.search(params);
-    }
-
-    @PreAuthorize(ROLE_ADMIN + OR + ROLE_BAM_MANAGER)
-    public void deleteCoverage(final Long bamId, final Integer step) throws IOException, ParseException {
-        coverageManager.delete(bamId, step);
     }
 }
